we can now edit behavior while movie is playing

+++ move_annotations_gui.py
@@ -42,7 +42,7 @@ app = QApplication(sys.argv)
 
 # sleep function
 def sleep(seconds):
-    QTest.qWait(seconds * 1000)
+    QTest.qWait(int(seconds * 1000))
 
 
 # named tuple for behavior
